reorganize icon files
