+++ index.html
@@ -3,7 +3,8 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="icon" href="src/assets/favicon.png" />
+    <link rel="icon" href="public/favicon.png" />
+    <link rel="apple-touch-icon" href="public/favicon.png" />
     <title>webmpm</title>
   </head>
   <body>
